New features
With this commit is now possible to start researches and collect its rewards from within the app.
It is also possible to open personal crates, even though, due to my lack of farm I don't have enough crates to test the possibility to open a X amount of crates instead of the fixed ones locked by UI on the site (1 and 5). If that's not possible I will address it in a further commit.
Furthermore on the personal crates opening, due to the fact that I don't know the structure of the various replies (timed versions, perma etc..) i just for now show only item and duration_type.